# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blood Donation Management System API",
        "description": "API documentation for Blood Donation Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/v1/donors": {
            "get": {
                "tags": [
                    "Donors"
                ],
                "summary": "Get list of all blood donors",
                "description": "Returns a list of donors with optional filtering by blood group.",
                "operationId": "7fa3a663d2963d8009389909f8b0de44",
                "parameters": [
                    {
                        "name": "blood_group",
                        "in": "query",
                        "description": "Filter donors by blood group (e.g., A+, B-, O+)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DonorResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Donors"
                ],
                "summary": "Register a new donor",
                "operationId": "c98958fba82450f50b33ec7b2a69860f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userId",
                                    "blood_group",
                                    "gender",
                                    "age"
                                ],
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "blood_group": {
                                        "type": "string",
                                        "example": "O+"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 26
                                    },
                                    "lastDonation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-01"
                                    },
                                    "isActive": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donor created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/donors/{id}": {
            "get": {
                "tags": [
                    "Donors"
                ],
                "summary": "Display a specific donor",
                "operationId": "7fcfd218097f395cc62cc0d44eb046c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of donor to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonorResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donor not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Donors"
                ],
                "summary": "Update an existing donor",
                "operationId": "3685e06a7c567a7f391df37bc428a446",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "blood_group": {
                                        "type": "string",
                                        "example": "A+"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 27
                                    },
                                    "lastDonation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donor updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonorResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Donors"
                ],
                "summary": "Delete a donor record",
                "operationId": "f7f05e9455a98923d305e16fd7000c7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donor deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Donor removed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/donors/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Donors"
                ],
                "summary": "Deactivate a donor",
                "description": "Disables the donor from being listed as available for donation.",
                "operationId": "63f0ff07f8a01414f5a52beb4856f84f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donor deactivated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/donors/{id}/activate": {
            "patch": {
                "tags": [
                    "Donors"
                ],
                "summary": "Activate a donor",
                "description": "Re-enables the donor for the active donation pool.",
                "operationId": "80c9091fbedd0eba0aa27b6a2b86aa32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donor activated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "DonorResource": {
                "title": "Donor Resource",
                "description": "Donor model schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "userId": {
                        "type": "integer",
                        "example": "5"
                    },
                    "blood_group": {
                        "type": "string",
                        "example": "O+"
                    },
                    "gender": {
                        "type": "string",
                        "example": "male"
                    },
                    "age": {
                        "type": "integer",
                        "example": 26
                    },
                    "lastDonation": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-01"
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}